Makefile: split out help targets in a modular way
When adding support for a new workflow I noticed it was easy
to forget to add respective targets for them. Fix this by making
these modular, and allowing a workflow to easily extend the top
level Makefile.

This also adds a few new Makefiles which will allow us to stuff
in work for each of those into them, removing clutter from the
top level Makefile.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>