Makefile: split out targets for extra_vars.yaml
The file extra_vars.yaml is generated from a series of input
variables. Most are just variables appended to ANSIBLE_EXTRA_ARGS.
But if you have some variables which are bit more complex we used
to just extend the target for the extra_vars.yaml file and augment
the commands for it there. This does not scale well and leaves
tons of clutter on the top level Makefile.

We modularize this by creating a variable with optional targets, and
each target workflow enabled can extend this as they see fit.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>