devconfig: install grub2 on debian
I've noticed that some installs sometimes ignore grub2, and
that's busted as GRUB_TIMEOUT is ignored. Some of this has
to do with the stupid cloud idea of always forcing the grub
timeout images by default with a stupid new file of
/etc/default/grub.d/15_timeout.cfg and forcing this down
everyone's throught. But we'll deal with that next.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>