devconfig: add a default vimrc
The most annoying thing is to ssh into a system to start hacking
and not have any sensible default settings to start hacking on Linux.
A most irritating default now on vim is the text highlighting feature
is disabled, fix that with the the set mouse=r setting.

Also use a default syntax color highlighting for default console
black backgrounds. We can later make this optional / let folks pick
whatever the hell they want.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>