kconfig: add support to fetch a defualt regcode using a script
Distributions such as SUSE Enterprise releases have a registration
code which is required if you want to make a plain intall useful.
Internal R&D teams may want to have the option to fetch an R&D
reg code from a server, which will check to see if the client is
indeed authorized to get it, if so it can provide it. Allow for
distributions to use their own custom and private script.

The way you'd support for this is you'd use kdevops as a git subtree,
and on top of it you can add the script, privately. Note that git
subtrees are different than git submodules. Git submodules suck.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>