Makefile: remove old substrees Makefile
We used to have our ansible roles as git subtrees. This provied
to be a maintainer's nightmare since it required tons of pushing
into different trees when you do an update. Because of this we
folded all that code here and its now maintained within kdevops.

This does *not* mean I'm opposed to using git subtrees on kdevops,
quite the contrary. But they also have a place. So for instance,
the kconfig code is a good candidate which should probably get
a git subtree upstream, and then we'd downstream consume it. This
is because we wouldn't update kconfig code as often. And when we
want to do that we should just have to do something like:

  make refresh

But right now this old Makefile can just confuse folks, so nuke it.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>