blktests: promote nbd a first class citizen package under blktests
The nbd git tree is needed for a few distributions. An example
is on SLE15-SP2 and above, where the nbd cliend and server are
not avaiable as official packages.

For those distibutions releases we must compile nbd. Take the
SUSE hacks we had on the blktests ansible role out from the
SUSE anbile file and generalize this to allow any
distributions to take advantage of this.

We also do this to allow us to modify the repository more
easily, which we'll do in a subsequent patch.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>