blktests: stick to using nbd version tag nbd-3.21 instead
When we need to compile nbd we were using the latest and greatest.
That proved to have issues. I then fixed at least one autogen.sh
issue, but that's not enough. I've reported the issue and am
moving on with life by just forcing the release to nbd-3.21.

This is only done on distributions which need to compile nbd
from source.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>