README.md: update docs to reflect no external ansible roles
The documentation was still a bit outdated hinting we used
external ansible roles. We used to support that by having all
of our ansible roles out of tree, but that easily became a
maintainer nightmare as we updated them regularly.

Adjust documentation to reflect all this and phase out any
docs which no longer apply. While a it also clearly document
and recommend folks to do a simple test run for their first
try of kdevops.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>