kconfigs/workflows/Kconfig.shared: removed shared software tools
Adding this complexity was a mistake for a few reasons:

  * In practice setting up postfix can be a nightmare
  * The watchdog tool is replaced now with systemd watchdog
    functionality and we've already added support for that more
    easily into the devconfig role.

And so experience is showing us to stuff extra packages we may
want to offer into the devconfig ansible role and we just let
folks enable / disable them. For instance, kdump would be next
as we support that now on suse / debian. Once fedora gets support
we can enable a new kconfig option for that.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>