kconfigs/Kconfig.kernel-ci: enable kernel-ci by default
Enable the kernel-ci targets by default. This will only be relevant
if a specific target supports such targets anyway. And so for example
'make help' won't list more targets avaiable unless your workflow
supports kernel-ci.

This makes it also easy for you to spot the watchdog options for
fstests and blktests without having to first enable kernel-ci at
the end of the menu and then go back to the fstests or blktests
menu to enable the respective fstests/blktests watchdog.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>