fstests/augment_expunge_list.py: add kotd support
Each time the kernel gets updated we have to go and add
a new symbolic link to the sles directory. This is silly,
because if we have KOTD enabled we know we just want to
test the latest and greatest and this means this is
representative of the latest issues.

So if KOTD is enabled use a shortcut directory for the
generic sles release name.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>