reboot-limit: fix how we handle counting reboots
Right now the test just does one reboot. This is silly.
We also unfortunately had conflated kernel-ci steady state
with the target max loop count to succeed, before we consider
the test as passing. Let's add a new kconfig entry to
specifically let us specify how many reboots are needed
before we consider the test to "pass". We then move all
the code which does th reboot into one task file and
include that however many reboot times is specified.
This accomplishes running the reboot task however many
times is specified.

The steady state the is kept as it is used in the other tests.
That is, if we set the max reboots 100, and steady state is
set to 100, steady state will be achieved once we reboot
successfully 10,000 times.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>