kernel-ci: provide a default for KERNEL_CI_DEFAULT_STEADY_STATE_GOAL
If a distribution does not enable HAVE_KERNEL_CI_CUSTOM_DEFAULTS we
get a complaint by kconfig on 'make menuconfig' when parsing the
kconfig that the value for the int for KERNEL_CI_DEFAULT_STEADY_STATE_GOAL
is not defined.

This is contrary to how things worth with bools, so for ints it seems
we have to provide a real default in case the HAVE_KERNEL_CI_CUSTOM_DEFAULTS
was not set.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>