fstests: add scripts to help cleanup expunge directories and files
After running a series of tests you may find that there are no
common all.txt files generated for a result. This adds two scripts
which lets you generate that file first with common expunges found
and then another script which can be used to then trim each existing
file to only expunges that are not common.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>