fstests/augment_expunge_list.py: add support for leap shortcut
When we process results from fstests we want to be able to share
expunges between SLES15-SP3 and OpenSUSE Leap 15.3. Everything is
working fine except that when we do have new issues we create a new
directory with expunges for the Leap 15.3 kernel. Fix that by adding
a special check when leap is used and refer to the opensuses-leap
expunge directory which is a symlink to the respective sles release.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>