blktests: remove autoconf-archive from sles15sp4
autoconf-archive is required to build nbd from the
latest version of git. We however use our own git
tree so that we can keep track of stable trees
which we know do build. The latest git version of
nbd requires autoconf-archive. However it is not
available on sles15sp4. So make sure we don't try
to install it there.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>