playbooks: add a new hypervisor tuning role
If you are setting up a kernel-ci using virtualization you should
enable a series of recommended kernel tuning parameters which can
help with memory on your hypervisor. We start with adding support
for KSM and zswap which in practice in SUSE's kernel-ci experience
has proven to save us about 135 GiB of RAM on a hypervisor running
about 60 guests.

This makes use of the LOCALHOST_SETUP_WORK variable on the top
Makefile, and by doing so, this ensures we always run this playbook
whenever 'make' is run. We want to do this prior to bringing guests
up, as otherwise ksm / zswap will have to catch up, and this can take
time, hours maybe. If we do it from the start we get the benefit from
it right away.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>