Initial import
This takes only the code needed from init-kconfig [0] which I think we
need to make this effort an independent git subtree. I'll try to then
see if I can get init-kconfig to use this tree as a git subtree.

If that works then I'll try to run the script to sync this tree
to the latest kconfig code upstrea:

scripts/update-upstream-kconfig.sh

[0] https://github.com/mcgrof/init-kconfig

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>