dbench: add a new role to compile dbench
dbench went missing on debian testing. There are tons of reasons for
this, but the main one is that this thing doesn't even compile anymore.

I've forked it and fixed the compilation issues. Use my tree for now
as I am not sure when when we'll get the maintainer to merge my fixes
and also to get debian testing to carry the package.

Since I can only confirm debian testing is affected at this point,
provide a role which enables the build only for debian. Other distros
can be added as folks wish if they need to. For now I'm not providing
any kconfig logic here, even to allow for custom git trees for dbench,
because this seems overkill at this point. If folks really need it
they can go ahead and extend this as I did with CONFIG_GIT_ALTERNATIVES
for different fstests git trees.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>