bootlinux: bump to next-20220211
Add a kernel configuration for linux-next next-20220211 and
update our bootlinux ansible role in Kconfig to use it
as the new default.

The next-20220209 and next-20220210 had a few issues which
would come up on some systems:

 * double declaration of binfmn_misc: crashes ppc but not x86_64
   This issue was fixed after I informed Stephen [0] to drop the
   extra fix from Andrew as Linus had already merged that fix.
 * fs/jbd2/journal.c triggers a kernel warning with kmem_cache_zalloc()
   usage. Paul reverted his addition of the WARN_ON_ONCE() later [1]

There is one more pending issue I can't wait for us to get added to
linux-next, a fix for zone block devices due to a regression on the
zone append command [2].

But we should at least now be able to run fstests for non zone block
devices on a clean kernel build.

[0] https://lkml.kernel.org/r/20220211184647.0a62dad4@canb.auug.org.au
[1] https://lore.kernel.org/linux-ext4/20220210091648.w5wie3llqri5kfw3@quack3.lan/T/#mfc3683a68aaec8595cd5f26e138453132a3e1d43
[2] https://lkml.kernel.org/20220211093425.43262-2-p.raghav@samsung.com

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>