vagrant: add initial zone storage support
This forces two new qemu drives when KVM is used which expose
2 NVMe ZNS drives. Note that this requires at least qemu 6.2.0.
You are expected to be using debian-testing or OpenSUSE Tumbleweed
as of today.

So you should not need to compile qemu on your own if using these
distributions. I can't say I'm going to care to support older
versions of qemu as kdevops stricly has always aimed at working
with rolling distributions.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>