bootlinux: add linux-5.17.y
The latest stable coming up is linux-5.17.y. Add support for it.
We're on rc7. Soon this will be baked. Let's start testing this
kernel and building a baseline for it.

We make it the default if the stable kernels are selected, but
you are free to pick another stable kernel.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>