Makefile: add support to limit hosts
You can use ansible-playbook -l to limit the hosts you want
to run on. We already use this for grouping hosts into baseline
and development. But it would be useful to limit the hosts
further without having to run ansible-playbook manually given
some playbooks like bootlinux have extra variables when running
them so constructing these manually is a bit difficult.

So provide a simple helper variable which can be used by
ansible playbooks to restrict running playbooks on specific
hosts further.

For example:

make linux HOSTS="kdevops-xfs-crc kdevops-xfs-reflink"

This will git clone linux, configure and install it on only
the hosts specified.

This feature is specially useful when building linux given that
if you are pushing the limits of a hypervisor you may end up
getting close to running out of memory after which libvirt will
just start shutting down guests and so eventually things will
start failing on those hosts which are shut down. If this
happens you have to just be aware you are running out of
memory and probably reconsider the number of guests you want
to work with. But still, you may want to run the above
command in other situations.

This is generally useful so just propagate support or it
through the different other important playbooks we support
for kernel development and testing:

 * bootlinux
 * fstests
 * blktests

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>