xfs: add an initial baseline for v5.10.105
This adds an initial baseline for v5.10.105.
Confidence in this baseline is just 1, meaning
that only 1 run of fstests has been made. I'll
let this spin a few times so to see if any new
issues are found through more loops on fstests.

A high degree confidence in a baseline is needed
given some tests may have a lower failure rate than
just 1/1. For example some tests may fail 1/40 or
worse 1/200, or even crazier 1/600. These *do*
happen.

If you don't create a baseline with a high degree
of confidence in failure rate it means a new failure
might be treated as a regression, but it may not be
if your confidence in the original baseline was not
high.

Note that I skipped testing the reatltime dev section,
that is because my home system ran out of memory to spawn
another guest. I'm trying to get more system resources
to run more sections. But this will do for now.

The all.txt represents failure found on all sections
tested, that is common failure for all sections, or
at least a few sections. Experience shows that if
a failure is happening on maybe 2 out of 6 sections
the failure rate could just be lower on the other
sections and testing fstests in a loop against them
will yield an eventual failure as well. To avoid a
failure like that interrupting establishing high
confidence in a baseline we strive to make a failure
common if its present in at least two sections.

Effort can later be done, *if needed*, to determine if a
failure is really generic to all sections or not, because
fixing an issue on one section may fix it for all. The
value in putting a failure on all.txt right away if
present in at least two sections is we want to create
a baseline fast.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>