fstests: fix /dev/loop sparse file descriptions
The itemized list of loopback files was off, fix this,
and adjust also the resulting SCRATCH_DEV_POOL set, it
had a duplicate of /dev/loop6. If is unclear if this
ever created a false positive, so far no tests have
caused one. To answer this question the question to ask
if which test requires 3 drives from SCRATCH_DEV_POOL
and we can check the existing expunge files.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>