vagrant: add kconfig for enabling ZNS
ZNS support was already added but it was forced down every user.
And these drives are only spawned when using libvirt but we were
not doing the work for Virtualbox yet.

Add Kconfig entries which adds the magic behind the logic here,
so to allow users to enable ZNS if they want it, and so to
express that virtualbox does not yet have ZNS support.

We do this as eventually we'll want to expand on the number of
ZNS drives we bring up when enabling ZNS so to enable ZNS filesystems
testing.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>