vagrant: bump number of ZNS drives to 14
Most of the work in the prior patch was so that you only bring up
these 14 NVMe ZNS drives if you really need them. We'll want them
for testing fstests as we need a lot of drives for testing filesystems
if we dream up the sky for ZNS. This brings parity to our loopback
strategy for testing with regular block devices.

This strategy should ensure that folk who don't enable ZNS don't
see tons of ZNS drives come up and that doesn't incur any cost
on memory on the guests. Folks who do want them are going to use
them.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>