fstests: add initial baseline for btrfs for v5.17-rc7
Three profiles / sections are supported right now.
This is the result of a test against all 3 profiles. The
expunges are not optimal in the sense that there are quite
a bit of failures shared between test sections, and so
we'll demonstrate how we can easily extract the common
failures and remove these later from the profiles where
common failures are annotated.

The current confidence in this baseline is 1: only one loop.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>