install_vagrant: enhance the redhat install
Make it clear our goal on redhat distros that our goal
is to force install the zip file directly. Force the fact
force_install_zip=true.

Likewise the manual installer does not verify and ensure
that you have a few packages to actually use vagrant properly,
so provide those packages as well.

Lastly, ensure we nuke the existing version of vagrant if
force_install_zip=true to ensure we actually upgrade in case
a user is stuck with an ancient release.

Note: for fedora we should strive to avoid this holy hell,
actually any sensible distribution will learn fast that this
path of manually installing vagrant-libvirt lies complete
madness and so the only thing sensible to do is to package
them.

I'll expand a bit more on this to help distributions when
evaluating if they should package vagrant or vagrant-libvirt:

If you are an enterprise release you probably don't because
of the entire holy hell dependency chain and the amount of work
you are then implying to support for many years all that ruby
crap. However you probably have some sort of "public" distro
oulet, so use that and package vagrant and vagrant-libvirt
crap there. The vagrant-libvirt stuff needs tons of work that
someone should just do to make vagrant libvirt provider a proper
first class citizen as the virtualbox one is, and so we don't have
to depend on two stupid packages. I think kdevops illustrates well
with manual commands what needs to get done to get this to happen.

Holy hell and can we simply have a rewrite of vagrant to some
sensible programming language other than this ruby crap? </rant>

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>