Vagrantfile: add support for a custom libvirt storage pool path
Prior to this change we always override the default for the
vagrant-libvirt plugin for the libvirt.storage_pool_path with:

libvirt.storage_pool_path = File.dirname(__FILE__)

__FILE__ in this case is kdevops Vagrantfile, so if you git cloned
kdevops in $HOME/kdevops this pool path is set to $HOME/kdevops/vagrant.
This is broken for a few reasons, the latest discovered reason being
is stated below which forces us to now change the default mechanism to
something more sensible:

1) If using libvirt vagrant tries to chown this directory to root,
   and for obvious reasons this is stupid. This means a user hacking
   on kdevops under $HOME/kdevops/ now sees $HOME/kdevops/vagrant owned
   as root. This presents all sorts of stupid issues requiring manual
   changes by the user.

2) If you git cloned kdevops under a directory under $HOME and if the
   $HOME directory has the permissions o-rw it means even if qemu runs
   as a user qemu, it won't be able to access the nvme drives or the
   images downloaded / created for the guest through kdevops.

Let's change the default to use something outside of the $HOME, and let
users manually select the old behavior. Since some Linux distributions
(SUSE) might have been relying on the prior behaviour so we enable the old
behaviour only for those distributions. We leave it up to those
distributions to revise and ensure that migrating to the new behaviour
is fine and folks using kdevops in production are ready for this change.

Since the pool path is also used on Virtualbox we refer to this as the
kdevops pool path. Virtualbox uses the pool path to create the additional
nvme drives only, so we create kconfig symbols for both libvirt and virtualbox
and provide descriptions of what this pool path means for each of them.

The last step is to ensure the pool path has the correct expected
permissions as otherwise we won't be able to get a user to read
or write to that directory. We take the liberty to change permissions
on the pool path as the user is aware of this during configuration,
however it would be stupid and wild to assume we can or should go
beyond that directory or files already on that directory. And this
is precisely why using anything based on $HOME was stupid to begin
with, although at first it probably it seemed like a nice idea.

This should fix the first 'make bringup' issues faced on a few
distributions which have o-rx on $HOME if a user git cloned kdevops
inside their $HOME somewhere, as our default for the kdevops pool
path is now is outside $HOME. If the user insists to use the pool
path within the $HOME directory at least there is enough documentation
to enable the to do what they need (chmod o+rw $HOME), but we're not
going to do that for them.

This lets us also simplfy out all the crazy ruby chown / mkdir / chgrp
stuff with just 2 nice ansible tasks delegated to the install_vagrant
ansible role. This is also another demonstration as to *why* the lease
vagrant does the better.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>