vagrant: add support for libvirt sessions
You can use libvirt either through the system or session URI.
The system URI assumes libvirt is run as root, whereas session
URIs have libvirtd instances run as a local user.

There are pros/cons for both, for details see:

https://wiki.libvirt.org/page/FAQ#What_is_the_difference_between_qemu:.2F.2F.2Fsystem_and_qemu:.2F.2F.2Fsession.3F_Which_one_should_I_use.3F

Using session support is required on Fedora as that is the default
there so this also adds full support for Fedora as a new distribution
we can claim support on kdevops.

Getting vagrant to use the session URI requires quite a bit of work
and it would seem not quite easy to grok. Details of random issues
are discussed here:

https://github.com/hashicorp/vagrant/issues/11020

Add initial support for libvirt session, and extend vagrant
to use it as well when this is enabled. We enable this only
right now by default if a user is on Fedora as that's where
this is seen as enabled by default now. I've tested debian-testing
and Ubuntu 21.10 and session support does not work there yet.
Hopefully the Kconfig entry for this feature is clear enough,
don't enable session support unless you are on Fedora. If you
enable it on other distributions you are on your own.

This is considered "initial support" as this at least ensures
we get 'make bringup' working well. We have 'make fstests',
'make blktests' and then 'make linux' target to test later.
It should in theory all work now though! Famous last words.

Note: if you relied on the additional interfaces with private IPs
added with system session as the default you can't use those now,
ie, they won't be created due to limitations of how using libvirt
through a session works. For details see the above or just this
entire commit.

With session support users can now just run virsh commands as a
regular user. Things like virsh list, virsh reset, virsh console,
etc.

While testing session support on Ubuntu I realized that the release
Ubuntu 21.10 also works so update the docs to relect that for the
hypervisor.

Reported-by: Josef Bacik <josef@toxicpanda.com>
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>