Kconfig: add symbols for your current distribution
We will want to add support to change defaults for kconfig
entries later depending on which Linux distribution you are
using, if you are using Linux. For example later on we will
want to change defaults if you are on Fedora as we want to use
libvirt with session mode. But we can't do this yet as we have
no way to know within the kconfig symbols when your local
distribution used is fedora.

This adds support for itemizing symbols for distributions used.
This will let us later use them in subsequent patches.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>