btrfs: remove common failures from btrfs_simple for v5.17-rc7
Likese btrfs_simple has no different failures so just remove
its own expunge file.

BTW I don't expect these chagnes to be itermized like this, this was
just an example to demonstrate the helper scripts which we should
eventually integrate into the fstests ansible role so that this
trimming and optimizations of the expounges is done as a post process
task on the host after a test.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>