vagrant: add a couple of extra non-power of 2 ZNS drives
The concept of whether or not we add support for non-power of 2 (NPO2)
zone size ZNS drives is currently an active topic. So in order
to help test those drives along with power of 2 zone sizes, add
a couple of NOPO2 ZNS drives. This should help kernel developers use
one single system to test / evaluate both worlds easily.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>