fstests: simplify oscheck's and gendisk.sh section inference
For a long time oscheck.sh was a kitchen sink for hacks, but
I wrote it before I adoped ansible. Now that ansible does all
the magic stuff for us we can simplify it further. For instance
oscheck.sh used to run for all sections which were on your
configuration. We no longer want that.

Likewise it always assumed statically your configuration file was
in configs/.config. But fstests uses its own ordering, so embrace
that.

So just infer the section using the hostname as we do with ansible,
and then allow the user to override the section. That's it.

Since we have a default section, it just means we must process
that first too with parse_config_section().

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>