fstests: standardize configurations for filesystems a bit more
If we want to use the [defaults] section it must not have things
which we need to override later, given fstests's check may simply
not override it for us. And so only put into [defaults] things
which are generic for all sections.

This fixes some odd issues I was seeing with ZNS drives where the
TEST_DEV changed from the ZNS drive to a local /dev/loop16, and
we mounted one over the other! Keep it simple then and force
full explicit configuration for each section. The top level
default just has the generic FS stuff for all the sections.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>