fstests: force PYTHONUNBUFFERED=1
The kdevops top level Makefile does export PYTHONUNBUFFERED=1
so to force ansible to not buffer output and to flush out
to stdout right away. This is to help the user know better what
task is being run. Sadly in practice I don't see this working,
so try to force this on the command line, and see / hope that
this is the expected use.

If this in turn *does* work as expected we can just modify
something lik $(Q) or something like that to have it.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>