install-deps: Update cache before debian install deps
Sometimes when we do a make destroy => make bringup => make fstests
skipping the make linux as it was already done before, it results in a
failure because of not doing apt update cache in install deps of fstests
or blktests.

This also makes it consistent with redhat where cache
is always updated before install.

Signed-off-by: Pankaj Raghav <p.raghav@samsung.com>
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>