fstests: add failures for btrfs_simple for v5.17-rc7
As we test ZNS it is also important to compare apples to apples.
So the test section btrfs_simple tests the same built filesystem
but without ZNS drives. These are a few failures with a failure
rate of 1/1, ie always fail.

Update the ZNS section to illustrate which are shared failaures.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>