xfs: add initial v5.17-rc7 baseline
This adds an initial baseline for xfs for v5.17-rc7.
The confidence in this baseline is just 1, meaning
'make fstests-baselin' was run just once. There are
common failures here but we'll use a script to first
extract those and then another to remove common failures
from each sections. We'll use a separate commit log
for that work to demonstrate how that's done.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>