kdevops: make sure loop is loaded before starting fstests
There's a weird race where we try to setup the loop devices and the loop
module doesn't load in time and we fail to create the first loop device.
This makes fstests fail because the initial loop device isn't created.
Fix this by pre-loading the loop device to make sure we can create
everything properly.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>