blktests: oscheck.sh: infer test group based on hostname
Follow the strategy used to test fstests with kdevop's
oscheck.sh to infer the group based on the hostname by
default if group was not specified on the command line.

The way devops run blktests is through ansiblwe by specifing
the group to test inferred by the hostname already. However
if one logs into he guest and tryies to run oscheck.sh
manually it will try to run all tests. Let's skip that
default mode and just stick to one group by default,
based on the hostname.

Also provide a sanity check to ensure that the group used
is something blktests expects.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>