bootlinux: bump to next-20220408
linux-next next-20220408 boots however be sure to have at least
qemu-system-x86_64 v6.2.0 because if you are using an older version
of qemu the UUID generated for the nvme drives may not be unique and
the new Linux kernel commit titled "nvme: check that EUI/GUID/UUID are
globally unique" seems to prevent older version of the nvme drives
attached to be spawned, leading to a failure at boot since we use
/dev/nvme0n1 for our /data/ partition.

qemu-system-x86_64 v6.2.0 still seems to generate something not so
random for all drives but it seems to be working... so this is still
being looked into.

No new changes to the kernel config is done yet. We'll do that next
to make the changes easier to read.

[0] https://lore.kernel.org/all/20220224192845.1097602-5-hch@lst.de/

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>