xfs: add two more failures for xfs on v5.10.105
generic/482 is failing on multiple sections, just with a
different failure rate per section. Just expunge the test for
all sections as we can assume this can likely fail for all
sections but with a lower failure rate. This will allow us
to move on with chasing our baseline faster.

xfs/060 is a new observed low failure rate failure so keep
that only for the section found so far.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>