btrfs: add generic/648 as failing for v5.17-rc7
It is odd to see generic/648 fail, it fails with a low failure
rate of about 1/50 on the btrfs_noraid56 section however when
inspecting the failure I can't see anything particular about
the failure which makes it seems its only specific to that
section. So expunge that test for all sections so we can move
on with life with the baseline. The failures can be observed
in the gist [0].

[0] https://gist.github.com/mcgrof/c3f6dae20800da6f1bda607d0c0275b3

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>