ext4: add new ext4_advanced_features section
This adds the section ext4_advanced_features for ext4.
It is taken from:

xfstests-bld/kvm-xfstests/test-appliance/files/root/fs/ext4/cfg/adv

The expunge is taken from its respective

xfstests-bld/kvm-xfstests/test-appliance/files/root/fs/ext4/cfg/adv.exclude

The *any* directory is used for this expunge as we follow
xfstest-bld full expunge for any kernel as well and we really have
not tested this section yet on v5.17-rc7. So the baseline for
v5.17-rc7 is empty right now for this section.

This comit serves as an example commit / documentation for how to
add a new section to any filesystem and how to generalize and expunge
file for that section for any kernel. This change alone allows you
enable/disable one more dedicated guest when testing ext4 and if
enabled, 'make fstests-baseline' will run tests with the configuration
supplied.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>