fstests/find-common-failures.sh: add support for a lazy baseline
Often times you may find a failure happens in only two sections with
a failure rate of 1, meaning it alwyas fails on them. However, this
typically is a sign that this test can also fail on other sections
easily but perhaps the failure rate is lower. Completing a baseline
takes time, and re-doing a full test just to discover this is
cumbersome. Add the option so that find-common-failures.sh can accept
a lazy parameter so that if an expunge is found at least twice in
two separate sections the expunge gets promoted as generic expunge
for all sections for the filesystem. This can help develop a baseline
faster.

If using a lazy baseline ensure that the new failures get a comment
to annotate that the test may have been found at least in only two
sections.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>