xfs: fix comment about xfsprogs regression on v5.10.105
I inspected the system where the baseline was last run, and it
was using xfsprogs 5.15.0-1+b1 so a regression due to using a new
xfsprogs is probably is not the reason. The only other thing left
to explain this is that the old baseline was using btrfs to place
the truncated files for the nvme drives used by qemu *and* also
for where the sparse files for the loopback files are used.

That is:

CONFIG_WORKFLOW_DATA_FSTYPE="btrfs"
CONFIG_FSTESTS_SPARSE_FSTYPE="btrfs"

Now the guests are using:

CONFIG_WORKFLOW_DATA_FSTYPE="xfs"
CONFIG_FSTESTS_SPARSE_FSTYPE="xfs"

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>