workflows/Kconfig: be consistent when enabling fstests or blktests
We have two kconfig variables which we use to be able to express
when we are going to enable fstests or blktests, either as a dedicated
set of tests or when we want to enable testing both fstests and blktests
in one system. But right now we only select this kconfig variable when
we are using a dedicated system. This is not an issue as the kconfig
is a kconfig symbols are bools which are set default to y if either
the test is dedicated or not.

But to be pedantic, and clear, let's make sure the tests select the
respective kconfig for each case as we'd expect to see it. Otherwise
this can confuse folks reading this.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>