playbooks: add a common playbook a git reset task for kdevops
Two playbooks share the concept of git cloning kdevops into
the target nodes (guests, cloud hosts, baremetal hosts) so that
expunge files can be used for avoiding tests. If you decide
you want to change the URL for that git tree it may not be
so obvious what to do.

Fortunately the solution is simple. You just tell ansible to use
the new git tree URL. That's it. It won't remove the old directory
and things work as expected.

But since we use the kdevops git tree on both fstests and blktests
it is not so obvious to developers that the thing to do here is
to just run 'make fstests' or 'make blktests' and even that is not
as efficient as that will also re-clone the fstests or blktests
tree respectively. When we just want to reset the kdevops git tree
we currently have no semantics to specify that. But since this is
a common post-deployment goal, just add a common playbook that let's
us do common tasks.

All we need then is the kconfig logic to define when some commmon
tasks might make sense. So to reset your kdevops git tree, all you
have to do now is change the configuration for it, then run:

make
make kdevops-git-reset

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>