kdevops: make linux-kdevops the default tree
At LSFMM 2022 at Palm Springs it was discussed that we should
*strive* towards a shared expunge list for fstests / blktests. Although
that effort requires splitting out the expunge list from kdevops to a
new git subtree, sharing a git tree for commit work for kdevops seems
also innevitable and desirable since some of us are already
collaborating on at least one shared test runner, kdevops.

So we can start by using a shared organization for what we need
to share, we call this organization linux-kdevops [0]. This encompasses a
few usual suspects git trees which can be used by both fstests and
blktests picking a "stable" sort of git sha1sum for each and always
striving towards the latest:

  * fstests
  * blktests
  * dbench
  * ndd

We'll be using the linux-kdevops organization for what trees we use and
trust that those in the organization will communicate what is needed before
making controversial changes.

This is perhaps the first controversial change, but it only applies to
kdevops, so if you are not using kdevops as a test runner you probably
won't care.

Those using kdevops should become aware that they should change
their default upstream to use linux-kdevops URL now:

So we switch from:
https://github.com/mcgrof/kdevops
To:
https://github.com/linux-kdevops/kdevops

Users which change this on a live environemtn would also then just have
to change the CONFIG_WORKFLOW_KDEVOPS_GIT on their configuration so
they'd just run:

 make menuconfig # set CONFIG_WORKFLOW_KDEVOPS_GIT to https://github.com/linux-kdevops/kdevops
 make
 make kdevops-git-reset

Those who have commit access to the organization linux-kdevops can
then just commit as needed to help move baselines forwards and if
and when something comes up which really seems controversial we can
use the mailing lists as with this change.

[0] https://github.com/orgs/linux-kdevops/

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>